Scope GoogleGossip CSS to only that component
The CSS contained global selectors like `input` that modified the style
for *all* other games. This should be addressed at a framework level (I
guess), but it's easy enough to fix case by case and is a good learning
opportunity for students.

GoogleGossip also incorrectly used `class` instead of `className` in
JSX. I took this opportunity to fix it.

Fixes #28